changed float to decimal and started on dtos

+++ template/Cryptocop.Software.API.Models/Entities/ShoppingCartItem.cs
@@ -5,6 +5,6 @@ public class ShoppingCartItem
     public int Id { get; set; }
     public int ShoppingCartId { get; set; }
     public string ProductIdentifier { get; set; } = null!;
-    public float Quantity { get; set; }
-    public float UnitPrice { get; set; }
+    public decimal Quantity { get; set; }
+    public decimal UnitPrice { get; set; }
 }
